Add support for weighted queues
Introdude shuffle strategy for when resque reserves a job. This allows for
random queue order and therefore for all queues to reserve time to complete
jobs without having to wait for the higher order ones to empty.

By default the feature is off, which means the default strategy is not
affected.

Example to enable:

Replace `QUEUES=high,medium,low` with `QUEUES=high=3,medium=2,low=1` in order
to allow `medium` and `low` queues to also perform but in a lower frequency,
without having to wait for `high` to empty.