Improve shuffling algorithm
The current version does not produce the right frequencies in which the
jobs should be performed.

The new version implements a more complicated algorithm which produces the
expected results.

Examples:

`QUEUES=high=1,medium=1,low=1` is expected to consume one job from each queue
every third time (33.3%).

`QUEUES=high=2,medium=1,low=1` is expected to give high ~50% chance
for being first and 25% for medium and low.

`QUEUES=high=3,medium=2,low=1` is expected to give high 50% chance
for being first, 33% for medium and 17% for low.

NOTE: For the last example the current version will produce 65% for high,
30% for medium and 5% for low.