Introduce BlacklistExceptionsRegistrar module
This commit introduces a new module for accessing a class variable
named `blacklist`. This class variable represents exceptions that
will not be pushed to the failed queue when a job fails with an
exception that is included in the blacklist.
Every class that includes this module can access/manipulate its
`blacklist` class variable.
Also it adds a method called `blacklisted_exception_raised?` which
checks if the raised exception belongs to the backend's blacklisted
exceptions.