Add support for suspending & killing jobs
Support for killing a type of the currently running jobs:

A thread will run on each worker checking for kill operations redis keys
describing the currently running jobs to kill.

If a kill operation matches the current job's class or smart id
then the worker will kill its child process.

A kill operation can be placed with:

  Resque.killall('<class>', '<method>')

----------------------------------------------------------------------

Support for suspending a type of jobs:

Suspending means that the resque worker will not run the jobs
matching the given criteria and will requeue the jobs to a
'suspended:<incident name>' new queue.

Suspend operations can be started with:

  Resque.start_suspending('incident name', '<class>', '<method>')

And stopped with:

  Resque.stop_suspending('incident name')

Suspended jobs can be requeued to a given queue with:

  Resque.requeue_suspended('incident name', queue)